+++ _source/atoms/messages/index.js
@@ -1,6 +1,6 @@
-import ErrorMessage from './error/ErrorMessageContainer';
-import SuccessMessage from './success/SuccessMessageContainer';
-import SuccessIllustration from './success/SuccessIllustrationContainer';
+import ErrorMessage from './error-message/ErrorMessageContainer';
+import SuccessMessage from './success-message/SuccessMessageContainer';
+import SuccessIllustration from './success-illustration/SuccessIllustrationContainer';
 
 export {
   ErrorMessage,
