+++ _source/atoms/messages/success-illustration/SuccessIllustration.js
