+++ _source/atoms/messages/success-illustration/SuccessIllustration.scss
