+++ _source/atoms/messages/success-illustration/SuccessIllustrationContainer.js
