+++ _source/templates/booky/Booky.scss
@@ -9,9 +9,9 @@
 @import "../../atoms/checkbox/Checkbox";
 @import "../../atoms/headline/Headline";
 @import "../../atoms/paragraph/Paragraph";
-@import "../../atoms/messages/error/ErrorMessage";
-@import "../../atoms/messages/success/SuccessMessage";
-@import "../../atoms/messages/success/SuccessIllustration";
+@import "../../atoms/messages/error-message/ErrorMessage";
+@import "../../atoms/messages/success-message/SuccessMessage";
+@import "../../atoms/messages/success-illustration/SuccessIllustration";
 @import "../../atoms/radio/Radio";
 @import "../../atoms/textarea/Textarea";
 @import "../../atoms/file-upload/FileUpload";
